src/app/main/config.py
import os
from dataclasses import dataclass

from dotenv import load_dotenv

load_dotenv()


@dataclass
class DbConfig:
    user: str
    password: str
    database: str
    host: str
    port: int
    driver: str = "postgresql+asyncpg"

    @staticmethod
    def from_env() -> "DbConfig":
        return DbConfig(
            user=os.getenv("POSTGRES_USER"),
            password=os.getenv("POSTGRES_PASSWORD"),
            database=os.getenv("POSTGRES_DB"),
            host=os.getenv("POSTGRES_HOST"),
            port=int(os.getenv("POSTGRES_PORT")),
        )

    def build_db_url(self) -> str:
        return f"{self.driver}://{self.user}:{self.password}@{self.host}:{self.port}/{self.database}"
