pyproject.toml
[build-system]
requires = ["setuptools==65.5.0"]

[tool.setuptools]
package-dir = { "" = "src" }

[project]
name = "SpyCatAgency"
version = "0.0.1"
authors = [
    { name = "Dmytro Kovalenko", email = "danrapongha@gmail.com"},
]
description = "Spy Cat Agency test assessment"
requires-python = ">=3.11"
dependencies = [
    "asyncpg==0.30.0",
    "sqlalchemy==2.0.36",
    "alembic==1.13.3",
    "typing-extensions==4.12.2",
    "python-dotenv==1.0.1",
    "fastapi==0.111.1",
    "uvicorn==0.30.1",
]

[project.optional-dependencies]
lint = [
  "mypy==1.10.1",
  "ruff==0.5.3",
  "isort==5.13.2"
]

[tool.mypy]
ignore_missing_imports = true

show_error_codes = true
warn_unused_configs = true
no_implicit_optional = true

warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_subclassing_any = true
# disallow_any_generics = true # maybe turn on in the future
local_partial_types = true


[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true
combine_as_imports = true
remove_redundant_aliases = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ['ALL']
fixable = [
  'Q000',
  'Q001',
  'COM812',
  'D400',
  'PT001',
]
ignore = [
  'D',
  'N804',
  'B008',
  'BLE001',
  'RUF009',
  'UP035',
  'TID252',
  'D104',
  'ANN',
  'SLF001',
  'ARG',
  'D100',
  'PLR0913',
  'TCH002',
  'ANN101',
  'FA100',
  'TRY003',
  'TRY201',
  'EM',
  'PERF203',
  'TCH001',
  'TD002',
  'PTH201',
  'RSE102',
  'I001',
  'COM812',
  'ISC001'
]
